backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
supabase==2.3.0
psycopg2-binary==2.9.9

# Caching
redis==5.0.1
fastapi-cache2[redis]==0.2.1
hiredis==2.2.3

# AI and ML
langchain==0.1.0
langchain-openai==0.0.5
openai==1.6.1
langchain-community==0.0.10

# HTTP and requests
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26
geopy==2.4.1
pytz==2023.3

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production server
gunicorn==21.2.0